+++ CHANGELOG.md
@@ -10,9 +10,14 @@ All notable changes to this project will be documented in this file.
 - Add `--list-products` argument to get a machine-readable (JSON) output of all products and their versions ([#45]).
 - Add `--completions nushell` support ([#46])
 
+## Changed
+
+- Update default `--registry` value to oci ([#49]).
+
 [#44]: https://github.com/stackabletech/image-tools/pull/44
 [#45]: https://github.com/stackabletech/image-tools/pull/45
 [#46]: https://github.com/stackabletech/image-tools/pull/46
+[#49]: https://github.com/stackabletech/image-tools/pull/49
 
 ## 0.0.13 - 2024-09-06